Moved the ReturnCodeCorrectness enum into its own file. Added timed-out/not-available icons to the ag_case_setup_result component.
