+++ src/components/project_view/submission_detail/return_code_correctness.ts
@@ -1,0 +1,6 @@
+export enum ReturnCodeCorrectness {
+    correct = 'Correct',
+    incorrect = 'Incorrect',
+    not_available = 'Not Available',
+    timed_out = 'Timed Out'
+}
