+++ tests/test_project_view/test_submission_detail/test_ag_command_result.ts
@@ -4,7 +4,7 @@ import * as ag_cli from 'ag-client-typescript';
 import * as sinon from 'sinon';
 
 import AGCommandResult from '@/components/project_view/submission_detail/ag_command_result.vue';
-import { ReturnCodeCorrectness } from '@/components/project_view/submission_detail/mutation_suite_result.vue';
+import { ReturnCodeCorrectness } from '@/components/project_view/submission_detail/return_code_correctness';
 
 import * as data_ut from '@/tests/data_utils';
